refactor(vector): ValidateConfig free function; move Config cluster into vtypes
