+++ vector/pq_hnsw_snapshot_test.go
@@ -244,7 +244,7 @@ func TestPQHNSWPersistADCSurvives(t *testing.T) {
 		MmapPath:      filepath.Join(dir, "vecs.dat"),
 		GraphMmapPath: filepath.Join(dir, "graph.dat"),
 	}
-	if err := cfg.Validate(); err != nil {
+	if err := ValidateConfig(cfg); err != nil {
 		t.Fatalf("PQ-HNSW + Persistent(mmap) config rejected: %v", err)
 	}
 	h, err := newHNSW(cfg)
