+++ vector/sq_test.go
@@ -365,7 +365,7 @@ func TestSQHNSWPersistSurvives(t *testing.T) {
 		MmapPath:      filepath.Join(dir, "vecs.dat"),
 		GraphMmapPath: filepath.Join(dir, "graph.dat"),
 	}
-	if err := cfg.Validate(); err != nil {
+	if err := ValidateConfig(cfg); err != nil {
 		t.Fatalf("QuantSQ + mmap config rejected: %v", err)
 	}
 	h, err := newHNSW(cfg)
